feat: initial experiments on XSP format
