+++ msixvc/src/lib.rs
@@ -1,4 +1,5 @@
 mod crypt;
+mod common;
 pub mod math;
 pub mod models;
 pub mod streaming;
