+++ msixvc/src/math.rs
@@ -1,4 +1,4 @@
-use crate::models::constants::{
+use crate::models::xvd::constants::{
     BLOCK_SIZE, DATA_BLOCKS_IN_LEVEL0_HASHTREE, DATA_BLOCKS_IN_LEVEL1_HASHTREE,
     DATA_BLOCKS_IN_LEVEL2_HASHTREE, DATA_BLOCKS_IN_LEVEL3_HASHTREE, HASH_ENTRIES_IN_PAGE,
     LEGACY_SECTOR_SIZE, PAGE_SIZE, PAGES_PER_BLOCK, SECTOR_SIZE,
