+++ msixvc/src/models.rs
@@ -1,9 +1,2 @@
-pub mod constants;
-pub mod enums;
-pub mod flags;
-pub mod structs;
-
-pub use constants::*;
-pub use enums::*;
-pub use flags::*;
-pub use structs::*;
+pub mod xvd;
+pub mod xsp;
