+++ msixvc/src/models/xsp/structs.rs
@@ -1,0 +1,8 @@
+pub mod parsed;
+pub mod raw;
+
+pub use parsed::*;
+use crate::common::*;
+
+
+impl_struct!(XspHeader);
