+++ msixvc/src/models/xvd/constants.rs
