+++ msixvc/src/models/xvd/enums.rs
