+++ msixvc/src/models/xvd/flags.rs
