+++ msixvc/src/models/xvd/structs/raw.rs
