+++ msixvc/src/utils/xsp.rs
@@ -1,0 +1,24 @@
+use tokio::{fs::OpenOptions, io::BufReader};
+use crate::common::*;
+use crate::models::xsp::XspHeader;
+
+pub struct XspFile {
+    header: XspHeader
+}
+impl XspFile {
+    pub async fn parse_file(path: String) -> Result<Self, Box<dyn std::error::Error>> {
+        let file = OpenOptions::new()
+            .read(true)
+            .open(path.clone())
+            .await
+            .expect("Unable to open file");
+
+        let mut file = BufReader::new(file); 
+
+        let header = read_struct!(XspHeader, file)?;
+
+        Ok(Self {
+            header
+        })
+    }
+}
