+++ xodus-cli/src/main.rs
@@ -1,6 +1,8 @@
 use clap::{Parser, Subcommand};
 use xodus::tokens::TokenManager;
 
+use crate::SubCommand::Xsp;
+
 mod commands;
 mod license;
 mod package;
@@ -8,13 +10,15 @@ mod webview;
 
 #[derive(Subcommand)]
 enum SubCommand {
+    #[command(about = "Download msixvc or xsp files fo given game")]
     Download {
         product: String,
         #[arg(short, long)]
         market: Option<String>,
-        #[arg(long, default_value_t = false)]
+        #[arg(long, default_value_t = false, help = "Display download URLs instead of downloading")]
         dry_run: bool,
     },
+    #[command(about = "Dump CIKs for use with XvdTool")]
     License {
         #[clap(help = "Content Id of a license")]
         content_id: String,
@@ -23,6 +27,7 @@ enum SubCommand {
         #[arg(short, long)]
         market: Option<String>,
     },
+    #[command(about = "Extract locally stored msixvc file")]
     Extract {
         path: String,
         destination: String,
@@ -30,6 +35,7 @@ enum SubCommand {
         market: Option<String>,
     },
     Login,
+    #[command(about = "Download and extract the game through streaming algorithm")]
     Streaming {
         source: String,
         destination: String,
@@ -44,6 +50,10 @@ enum SubCommand {
         #[arg(short, long)]
         market: Option<String>,
     },
+    #[command(about = "Debug info on XSP file")]
+    Xsp {
+        file: String,
+    }
 }
 
 #[derive(Parser)]
@@ -121,6 +131,9 @@ async fn main() {
                 market,
             )
             .await;
+        },
+        Xsp { file } => {
+            commands::xsp::run(file).await;
         }
     }
